refactor: put remaining common code to nested::util
